Separate out the settings.
